refactor(test): split file-info tests by classifier extension license and naming concerns
